Refactor: split core ops #647
