+++ test/handlers/tool-configs/universal/core-operations-crud.test.ts
@@ -10,7 +10,7 @@ import {
   createRecordConfig,
   updateRecordConfig,
   deleteRecordConfig,
-} from '../../../../src/handlers/tool-configs/universal/core-operations.js';
+} from '../../../../src/handlers/tool-configs/universal/core/index.js';
 
 // Import types
 import {
