+++ test/handlers/tool-configs/universal/core-operations-search.test.ts
@@ -13,7 +13,7 @@ import {
 import {
   searchRecordsConfig,
   getRecordDetailsConfig,
-} from '../../../../src/handlers/tool-configs/universal/core-operations.js';
+} from '../../../../src/handlers/tool-configs/universal/core/index.js';
 
 // Import types
 import {
